Moved parsing of `#[sabi(impl_InterfaceType())]` to its own top-level module.
Added supertrait implication,
where (as an example) using `#[sabi(impl_InterfaceType(Ord))]` is equivalent to
using `#[sabi(impl_InterfaceType(Ord,PartialOrd,PartialEq,Eq))]`.
