+++ abi_stable_derive_lib/src/attribute_parsing.rs
@@ -1,5 +1,7 @@
-use syn::{Meta, NestedMeta};
 
+use syn::{
+    Meta, NestedMeta,
+};
 
 
 
@@ -27,5 +29,19 @@ where
         }
     }
 }
+
+
+
+///////////////////////////////////////////////////////////////////////////////
+
+
+
+
+
+
+///////////////////////////////////////////////////////////////////////////////
+
+
+
